optimized CPX imports and code quality
- export all CPX files through one cpx export
- deleted unnecessary widgets_overlay.dart and moved code directly into cpx_research.dart
- sort and refactor models, utils and widgets including imports
